ThemeProvider - Move component to `/theme`.
