+++ src/index.js
@@ -1,22 +1,5 @@
 // theme
-import React from 'react'
-import PropTypes from 'prop-types'
-import { ThemeProvider as MUIThemeProvider } from '@material-ui/core/styles'
-import DefaultTheme from './theme'
-
-export { DefaultTheme }
-export const ThemeProvider = ({ children, theme = DefaultTheme }) => {
-  return (
-    <MUIThemeProvider theme={theme}>
-      {children}
-    </MUIThemeProvider>
-  )
-}
-
-ThemeProvider.propTypes = {
-  children: PropTypes.node,
-  theme: PropTypes.object,
-}
+export { DefaultTheme, ThemeProvider } from './theme'
 
 // components
 export { default as Alert } from './alert'
